atualizando excptions 204 --> 404, update e delete --> 27/08/25 14:25

+++ src/main/java/com/devFelipe/basic_spring_project/services/UserService.java
@@ -4,10 +4,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.devFelipe.basic_spring_project.entities.User;
 import com.devFelipe.basic_spring_project.repositories.UserRepository;
+import com.devFelipe.basic_spring_project.services.exceptions.DatabaseException;
+import com.devFelipe.basic_spring_project.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -21,7 +27,7 @@ public class UserService {
 
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
-		return obj.get();
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 
 	public User insert(User obj) {
@@ -29,15 +35,43 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			// repository.deleteById(id);
+
+			if (!repository.existsById(id))
+				throw new ResourceNotFoundException("ID NOT FOUND");
+			repository.deleteById(id);
+
+		} catch (EmptyResultDataAccessException e) {
+			// EmptyResultDataAccessException
+//			System.out.println(e.getMessage());
+//
+//			e.printStackTrace();
+//			e.getMessage();
+			throw new ResourceNotFoundException(id);
+
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+
+		} catch (RuntimeException e) {
+			e.printStackTrace();
+			throw new ResourceNotFoundException(id);
+		}
 	}
 
 	public User update(Long id, User obj) {
-		User entity = repository.getReferenceById(id);
-		updateData(entity, obj);
-		return repository.save(entity);
+		try {
+			User entity = repository.getReferenceById(id);
+			updateData(entity, obj);
+			return repository.save(entity);
+
+		} catch (EntityNotFoundException e) {
+			e.printStackTrace();
+			throw new ResourceNotFoundException(id);
+		}
+
 	}
-	
+
 	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
